requirements.txt
APScheduler==3.6.3
asyncio==3.4.3
cachetools==4.2.2
certifi==2022.6.15
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.39
tornado==6.2
tzdata==2022.1
tzlocal==4.2
